conditions refactoring, fixing YAMLLinter
