+++ Content.Server/AlertLevel/AlertLevelComponent.cs
@@ -1,6 +1,6 @@
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype;
 
-namespace Content.Shared.AlertLevel;
+namespace Content.Server.AlertLevel;
 
 /// <summary>
 /// Alert level component. This is the component given to a station to
