+++ Content.Server/AlertLevel/Commands/SetAlertLevelCommand.cs
@@ -2,7 +2,6 @@
 using Content.Server.Administration;
 using Content.Server.Station.Systems;
 using Content.Shared.Administration;
-using Content.Shared.AlertLevel;
 using JetBrains.Annotations;
 using Robust.Shared.Console;
